Optimized console and reduced renders
