+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "echo-search",
   "productName": "echo-search",
-  "version": "2.4.3",
+  "version": "2.4.4",
   "description": "An application to search and replace strings in files recursively",
   "main": ".webpack/main",
   "scripts": {
