+++ src/Client/Components/Output.js
@@ -1,27 +1,17 @@
 import React, { useState, useEffect, useRef } from "react";
 import Box from "@mui/material/Box";
 import LinearProgress from "@mui/material/LinearProgress";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemText from "@mui/material/ListItemText";
 import Typography from "@mui/material/Typography";
 import Divider from "@mui/material/Divider";
-import IconButton from "@mui/material/IconButton";
-import Tooltip from "@mui/material/Tooltip";
-import FileIcon from "@mui/icons-material/InsertDriveFile";
-import FolderIcon from "@mui/icons-material/Folder";
+import Console from "./Console";
 
-const { ipcListen, ipcSend } = window.api;
+const { ipcListen } = window.api;
 
 const FAILED_SEARCH_HEADING = "Operation was not completely successful.";
 const MESSAGE_MODES = {
   success: "success.main",
   error: "error.main",
   info: "text.primary",
-};
-const MESSAGE_PREFIX = {
-  update: "Updated: ",
-  match: "Matched: ",
 };
 
 const getMessage = (msg) => {
@@ -52,26 +42,12 @@ const getMessage = (msg) => {
   };
 };
 
-const openFile = (msg, inFolder) => {
-  let filePath;
-  const request = inFolder ? "open:folder" : "open:file";
-  if (msg.includes(MESSAGE_PREFIX.update)) {
-    filePath = msg.replace(MESSAGE_PREFIX.update, "");
-  } else if (msg.includes(MESSAGE_PREFIX.match)) {
-    filePath = msg.replace(MESSAGE_PREFIX.match, "");
-  } else {
-    filePath = msg;
-  }
-  ipcSend(request, { filePath });
-};
-
 function Output({ isRunning }) {
   // due to async nature of useState, we might miss some messages, using ref
   const allMessages = useRef([]);
   const [hasError, setHasError] = useState(false);
   const [progress, setProgress] = useState(0);
   const [heading, setHeading] = useState("Searching for files...");
-  const [messages, setMessages] = React.useState([]);
 
   useEffect(() => {
     const showProgress = (update) => {
@@ -79,7 +55,6 @@ function Output({ isRunning }) {
       if (update.message) {
         const message = getMessage(update);
         allMessages.current.push(message);
-        setMessages([...allMessages.current]);
       }
       if (update.mode === "error" && update.error) {
         console.error(update.error);
@@ -102,7 +77,6 @@ function Output({ isRunning }) {
       allMessages.current.push(message);
       setHeading("All Done!");
       setProgress(100);
-      setMessages([...allMessages.current]);
     };
     return ipcListen("search:complete", onComplete);
   }, []);
@@ -113,7 +87,6 @@ function Output({ isRunning }) {
       allMessages.current.push(message);
       setProgress(100);
       setHasError(true);
-      setMessages([...allMessages.current]);
     };
     return ipcListen("search:fail", showError);
   }, []);
@@ -142,54 +115,7 @@ function Output({ isRunning }) {
       </Typography>
       <Divider />
       <Box sx={{ width: "100%" }}>
-        <List
-          dense={true}
-          sx={{
-            maxHeight: 300,
-            overflow: "auto",
-            pt: 2,
-          }}
-        >
-          {messages.map((msg, i) => (
-            <ListItem key={i}>
-              <Tooltip
-                placement="bottom-start"
-                arrow
-                title={
-                  msg.isFile && (
-                    <>
-                      <Tooltip title="Open folder in the file manager.">
-                        <IconButton
-                          size="small"
-                          onClick={() => openFile(msg.message, true)}
-                        >
-                          <FolderIcon fontSize="8" />
-                        </IconButton>
-                      </Tooltip>
-                      <Tooltip title="Open file in the desktop's default manner. Similar to double clicking on the file.">
-                        <IconButton
-                          size="small"
-                          onClick={() => openFile(msg.message)}
-                        >
-                          <FileIcon fontSize="8" />
-                        </IconButton>
-                      </Tooltip>
-                    </>
-                  )
-                }
-              >
-                <ListItemText
-                  sx={{
-                    color: msg.mode,
-                    wordBreak: "break-all",
-                    cursor: "default",
-                  }}
-                  primary={msg.message}
-                />
-              </Tooltip>
-            </ListItem>
-          ))}
-        </List>
+        <Console  messagesRef={allMessages}/>
       </Box>
     </Box>
   );
